Call restoreSnapshot(), finishSnapshotRestore() and completeEdge(), untested

+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/SnapshotContext.java
@@ -57,7 +57,7 @@ public class SnapshotContext {
     }
 
     void initTaskletCount(int count) {
-        assert this.taskletCount == Long.MIN_VALUE;
+        assert this.taskletCount == Integer.MIN_VALUE;
         this.taskletCount = count;
     }
